distributed V2 value objects to proper files
