+++ dsp-shared/src/main/scala/dsp/valueobjects/Iri.scala
@@ -5,6 +5,8 @@
 
 package dsp.valueobjects
 
+import com.google.gwt.safehtml.shared.UriUtils.encodeAllowEscapes
+import org.apache.commons.lang3.StringUtils
 import org.apache.commons.validator.routines.UrlValidator
 import zio.json.JsonDecoder
 import zio.json.JsonEncoder
@@ -19,6 +21,7 @@ sealed trait Iri {
   val value: String
 }
 object Iri {
+  type IRI = String
 
   // A validator for URLs
   val urlValidator =
@@ -26,6 +29,17 @@ object Iri {
       Array("http", "https"),       // valid URL schemes
       UrlValidator.ALLOW_LOCAL_URLS // local URLs are URL-encoded IRIs as part of the whole URL
     )
+
+  object KnoraInternal {
+    // The start and end of an internal Knora ontology IRI.
+    val InternalOntologyStart = "http://www.knora.org/ontology"
+  }
+
+  val KnoraAdminOntologyLabel: String     = "knora-admin"
+  val KnoraAdminOntologyIri: IRI          = KnoraInternal.InternalOntologyStart + "/" + KnoraAdminOntologyLabel
+  val KnoraAdminPrefixExpansion: IRI      = KnoraAdminOntologyIri + "#"
+  val SystemProject: IRI                  = KnoraAdminPrefixExpansion + "SystemProject"
+  val DefaultSharedOntologiesProject: IRI = KnoraAdminPrefixExpansion + "DefaultSharedOntologiesProject"
 
   /**
    * Returns `true` if a string is an IRI.
@@ -37,6 +51,157 @@ object Iri {
     urlValidator.isValid(s)
 
   /**
+   * Returns `true` if an IRI string looks like a Knora list IRI.
+   *
+   * @param iri the IRI to be checked.
+   */
+  def isKnoraListIriStr(iri: IRI): Boolean =
+    Iri.isIri(iri) && iri.startsWith("http://rdfh.ch/lists/")
+
+  /**
+   * Returns `true` if an IRI string looks like a Knora role IRI.
+   *
+   * @param iri the IRI to be checked.
+   */
+  def isKnoraRoleIriStr(iri: IRI): Boolean =
+    Iri.isIri(iri) && iri.startsWith("http://rdfh.ch/roles/")
+
+  /**
+   * Returns `true` if an IRI string looks like a Knora user IRI.
+   *
+   * @param iri the IRI to be checked.
+   */
+  def isKnoraUserIriStr(iri: IRI): Boolean =
+    Iri.isIri(iri) && iri.startsWith("http://rdfh.ch/users/")
+
+  /**
+   * Returns `true` if an IRI string looks like a Knora group IRI.
+   *
+   * @param iri the IRI to be checked.
+   */
+  def isKnoraGroupIriStr(iri: IRI): Boolean =
+    Iri.isIri(iri) && iri.startsWith("http://rdfh.ch/groups/")
+
+  /**
+   * Returns `true` if an IRI string looks like a Knora project IRI
+   *
+   * @param iri the IRI to be checked.
+   */
+  def isKnoraProjectIriStr(iri: IRI): Boolean =
+    (iri.startsWith("http://rdfh.ch/projects/") || isKnoraBuiltInProjectIriStr(iri))
+
+  /**
+   * Returns `true` if an IRI string looks like a Knora built-in IRI:
+   *  - http://www.knora.org/ontology/knora-admin#SystemProject
+   *  - http://www.knora.org/ontology/knora-admin#SharedOntologiesProject
+   *
+   * @param iri the IRI to be checked.
+   */
+  def isKnoraBuiltInProjectIriStr(iri: IRI): Boolean = {
+
+    val builtInProjects = Seq(
+      SystemProject,
+      DefaultSharedOntologiesProject
+    )
+
+    Iri.isIri(iri) && builtInProjects.contains(iri)
+  }
+
+  // Characters that are escaped in strings that will be used in SPARQL.
+  private val SparqlEscapeInput = Array(
+    "\\",
+    "\"",
+    "'",
+    "\t",
+    "\n"
+  )
+
+  // Escaped characters as they are used in SPARQL.
+  private val SparqlEscapeOutput = Array(
+    "\\\\",
+    "\\\"",
+    "\\'",
+    "\\t",
+    "\\n"
+  )
+
+  /**
+   * Makes a string safe to be entered in the triplestore by escaping special chars.
+   *
+   * If the param `revert` is set to `true`, the string is unescaped.
+   *
+   * @param s        a string.
+   * @param errorFun a function that throws an exception. It will be called if the string is empty or contains
+   *                 a carriage return (`\r`).
+   * @return the same string, escaped or unescaped as requested.
+   */
+  def toSparqlEncodedString(s: String, errorFun: => Nothing): String = {
+// Findings about this method:
+// - there is more cases to handle according to docs (https://www.w3.org/TR/rdf-sparql-query#grammarEscapes) - we use 1.1 version right?
+// - `'` doesn't appear on that list, but this method escapes it
+// - why `\r` throws error instead of being escaped?
+// - fun fact is that if I remove  StringUtils.replaceEach, for example `\t` passes unescaped, why?
+
+    if (s.isEmpty || s.contains("\r")) errorFun
+
+    // http://www.morelab.deusto.es/code_injection/
+
+    StringUtils.replaceEach(
+      s,
+      SparqlEscapeInput,
+      SparqlEscapeOutput
+    )
+  }
+
+  /**
+   * Checks that a string represents a valid IRI. Also encodes the IRI, preserving existing %-escapes.
+   *
+   * @param s        the string to be checked.
+   * @param errorFun a function that throws an exception. It will be called if the string does not represent a valid
+   *                 IRI.
+   * @return the same string.
+   */
+  def validateAndEscapeIri(s: String, errorFun: => Nothing): IRI = {
+    val urlEncodedStr = encodeAllowEscapes(s)
+
+    if (Iri.urlValidator.isValid(urlEncodedStr)) {
+      urlEncodedStr
+    } else {
+      errorFun
+    }
+  }
+
+  /**
+   * Check that the supplied IRI represents a valid project IRI.
+   *
+   * @param iri      the string to be checked.
+   * @param errorFun a function that throws an exception. It will be called if the string does not represent a valid
+   *                 project IRI.
+   * @return the same string but escaped.
+   */
+  def validateAndEscapeProjectIri(iri: IRI, errorFun: => Nothing): IRI =
+    if (isKnoraProjectIriStr(iri)) {
+      toSparqlEncodedString(iri, errorFun)
+    } else {
+      errorFun
+    }
+
+  /**
+   * Check that the supplied IRI represents a valid user IRI.
+   *
+   * @param iri      the string to be checked.
+   * @param errorFun a function that throws an exception. It will be called if the string does not represent a valid
+   *                 user IRI.
+   * @return the same string but escaped.
+   */
+  def validateAndEscapeUserIri(iri: IRI, errorFun: => Nothing): String =
+    if (isKnoraUserIriStr(iri)) {
+      toSparqlEncodedString(iri, errorFun)
+    } else {
+      errorFun
+    }
+
+  /**
    * GroupIri value object.
    */
   sealed abstract case class GroupIri private (value: String) extends Iri
@@ -45,15 +210,15 @@ object Iri {
       if (value.isEmpty) {
         Validation.fail(BadRequestException(IriErrorMessages.GroupIriMissing))
       } else {
-        val isUuid: Boolean = V2UuidValidation.hasUuidLength(value.split("/").last)
+        val isUuid: Boolean = Uuid.hasUuidLength(value.split("/").last)
 
-        if (!V2IriValidation.isKnoraGroupIriStr(value)) {
+        if (!isKnoraGroupIriStr(value)) {
           Validation.fail(BadRequestException(IriErrorMessages.GroupIriInvalid))
-        } else if (isUuid && !V2UuidValidation.isUuidSupported(value)) {
+        } else if (isUuid && !Uuid.isUuidSupported(value)) {
           Validation.fail(BadRequestException(IriErrorMessages.UuidVersionInvalid))
         } else {
           val validatedValue = Validation(
-            V2IriValidation.validateAndEscapeIri(value, throw BadRequestException(IriErrorMessages.GroupIriInvalid))
+            validateAndEscapeIri(value, throw BadRequestException(IriErrorMessages.GroupIriInvalid))
           )
 
           validatedValue.map(new GroupIri(_) {})
@@ -76,15 +241,15 @@ object Iri {
       if (value.isEmpty) {
         Validation.fail(BadRequestException(IriErrorMessages.ListIriMissing))
       } else {
-        val isUuid: Boolean = V2UuidValidation.hasUuidLength(value.split("/").last)
+        val isUuid: Boolean = Uuid.hasUuidLength(value.split("/").last)
 
-        if (!V2IriValidation.isKnoraListIriStr(value)) {
+        if (isKnoraListIriStr(value)) {
           Validation.fail(BadRequestException(IriErrorMessages.ListIriInvalid))
-        } else if (isUuid && !V2UuidValidation.isUuidSupported(value)) {
+        } else if (isUuid && !Uuid.isUuidSupported(value)) {
           Validation.fail(BadRequestException(IriErrorMessages.UuidVersionInvalid))
         } else {
           val validatedValue = Validation(
-            V2IriValidation.validateAndEscapeIri(
+            validateAndEscapeIri(
               value,
               throw BadRequestException(IriErrorMessages.ListIriInvalid)
             )
@@ -116,15 +281,15 @@ object Iri {
       if (value.isEmpty) {
         Validation.fail(ValidationException(IriErrorMessages.ProjectIriMissing))
       } else {
-        val isUuid: Boolean = V2UuidValidation.hasUuidLength(value.split("/").last)
+        val isUuid: Boolean = Uuid.hasUuidLength(value.split("/").last)
 
-        if (!V2IriValidation.isKnoraProjectIriStr(value)) {
+        if (!isKnoraProjectIriStr(value)) {
           Validation.fail(ValidationException(IriErrorMessages.ProjectIriInvalid))
-        } else if (isUuid && !V2UuidValidation.isUuidSupported(value)) {
+        } else if (isUuid && !Uuid.isUuidSupported(value)) {
           Validation.fail(ValidationException(IriErrorMessages.UuidVersionInvalid))
         } else {
           val eitherValue = Try(
-            V2IriValidation.validateAndEscapeProjectIri(
+            validateAndEscapeProjectIri(
               value,
               throw ValidationException(IriErrorMessages.ProjectIriInvalid)
             )
@@ -153,9 +318,9 @@ object Iri {
     def make(value: String): Validation[ValidationException, Base64Uuid] =
       if (value.isEmpty) {
         Validation.fail(ValidationException(IriErrorMessages.UuidMissing))
-      } else if (!V2UuidValidation.hasUuidLength(value)) {
+      } else if (!Uuid.hasUuidLength(value)) {
         Validation.fail(ValidationException(IriErrorMessages.UuidInvalid(value)))
-      } else if (!V2UuidValidation.isUuidSupported(value)) {
+      } else if (!Uuid.isUuidSupported(value)) {
         Validation.fail(ValidationException(IriErrorMessages.UuidVersionInvalid))
       } else Validation.succeed(new Base64Uuid(value) {})
   }
@@ -169,15 +334,15 @@ object Iri {
       if (value.isEmpty) {
         Validation.fail(BadRequestException(IriErrorMessages.RoleIriMissing))
       } else {
-        val isUuid: Boolean = V2UuidValidation.hasUuidLength(value.split("/").last)
+        val isUuid: Boolean = Uuid.hasUuidLength(value.split("/").last)
 
-        if (!V2IriValidation.isKnoraRoleIriStr(value)) {
+        if (!isKnoraRoleIriStr(value)) {
           Validation.fail(BadRequestException(IriErrorMessages.RoleIriInvalid(value)))
-        } else if (isUuid && !V2UuidValidation.isUuidSupported(value)) {
+        } else if (isUuid && !Uuid.isUuidSupported(value)) {
           Validation.fail(BadRequestException(IriErrorMessages.UuidVersionInvalid))
         } else {
           val validatedValue = Validation(
-            V2IriValidation.validateAndEscapeIri(
+            validateAndEscapeIri(
               value,
               throw BadRequestException(IriErrorMessages.RoleIriInvalid(value))
             )
@@ -202,15 +367,15 @@ object Iri {
       if (value.isEmpty) {
         Validation.fail(BadRequestException(IriErrorMessages.UserIriMissing))
       } else {
-        val isUuid: Boolean = V2UuidValidation.hasUuidLength(value.split("/").last)
+        val isUuid: Boolean = Uuid.hasUuidLength(value.split("/").last)
 
-        if (!V2IriValidation.isKnoraUserIriStr(value)) {
+        if (!isKnoraUserIriStr(value)) {
           Validation.fail(BadRequestException(IriErrorMessages.UserIriInvalid(value)))
-        } else if (isUuid && !V2UuidValidation.isUuidSupported(value)) {
+        } else if (isUuid && !Uuid.isUuidSupported(value)) {
           Validation.fail(BadRequestException(IriErrorMessages.UuidVersionInvalid))
         } else {
           val validatedValue = Validation(
-            V2IriValidation.validateAndEscapeUserIri(
+            validateAndEscapeUserIri(
               value,
               throw BadRequestException(IriErrorMessages.UserIriInvalid(value))
             )
