+++ dsp-shared/src/main/scala/dsp/valueobjects/Project.scala
@@ -14,11 +14,59 @@ import dsp.errors.ValidationException
 
 object Project {
   // A regex sub-pattern for project IDs, which must consist of 4 hexadecimal digits.
-  private val ProjectIDPattern: String =
-    """\p{XDigit}{4,4}"""
+  private val ProjectIDPattern: String = """\p{XDigit}{4,4}"""
 
   // A regex for matching a string containing the project ID.
   private val ProjectIDRegex: Regex = ("^" + ProjectIDPattern + "$").r
+
+  /**
+   * Given the project shortcode, checks if it is in a valid format, and converts it to upper case.
+   *
+   * @param shortcode the project's shortcode.
+   * @return the shortcode in upper case.
+   */
+  def validateProjectShortCode(shortCode: String, errorFun: => Nothing): String =
+    ProjectIDRegex.findFirstIn(shortCode.toUpperCase) match {
+      case Some(value) => value
+      case None        => errorFun
+    }
+
+  // A regex sub-pattern for ontology prefix labels and local entity names. According to
+  // <https://www.w3.org/TR/turtle/#prefixed-name>, a prefix label in Turtle must be a valid XML NCName
+  // <https://www.w3.org/TR/1999/REC-xml-names-19990114/#NT-NCName>. Knora also requires a local entity name to
+  // be an XML NCName.
+  private val NCNamePattern: String = """[\p{L}_][\p{L}0-9_.-]*"""
+
+  // A regex for matching a string containing only an ontology prefix label or a local entity name.
+  private val NCNameRegex: Regex = ("^" + NCNamePattern + "$").r
+
+  // A regex sub-pattern matching the random IDs generated by KnoraIdUtil, which are Base64-encoded
+  // using the "URL and Filename safe" Base 64 alphabet, without padding, as specified in Table 2 of
+  // RFC 4648.
+  private val Base64UrlPattern = "[A-Za-z0-9_-]+"
+
+  private val Base64UrlPatternRegex: Regex = ("^" + Base64UrlPattern + "$").r
+
+  /**
+   * Check that the string represents a valid project shortname.
+   *
+   * @param value    the string to be checked.
+   * @param errorFun a function that throws an exception. It will be called if the string does not represent a valid
+   *                 project shortname.
+   * @return the same string.
+   */
+  def validateAndEscapeProjectShortname(shortname: String, errorFun: => Nothing): String = {
+    // Check that shortname matches NCName pattern
+    val ncNameMatch = NCNameRegex.findFirstIn(shortname) match {
+      case Some(value) => value
+      case None        => errorFun
+    }
+    // Check that shortname is URL safe
+    Base64UrlPatternRegex.findFirstIn(ncNameMatch) match {
+      case Some(shortname) => Iri.toSparqlEncodedString(shortname, errorFun)
+      case None            => errorFun
+    }
+  }
 
   // TODO-mpro: longname, description, keywords, logo are missing enhanced validation
 
@@ -66,7 +114,7 @@ object Project {
         Validation.fail(ValidationException(ProjectErrorMessages.ShortNameMissing))
       } else {
         val validatedValue = Validation(
-          V2ProjectIriValidation.validateAndEscapeProjectShortname(
+          validateAndEscapeProjectShortname(
             value,
             throw ValidationException(ProjectErrorMessages.ShortNameInvalid(value))
           )
